Refactor driver interface code and update tracer
